+++ bin/clean.js
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 import * as fs from 'fs-extra'
 
 fs.emptyDirSync('./dist')